Refactor AutoEncoder + some doc (#80)
